Move `FinancialsMixin` to "strategies" directory
